refactored home page, changed sidebar style
