+++ lib/sidebar/sidebar.dart
@@ -1,215 +1,100 @@
-import 'dart:async';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:rxdart/rxdart.dart';
 import '../bloc.navigation_bloc/navigation_bloc.dart';
-import '../sidebar/menu_item.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
-class SideBar extends StatefulWidget {
+class SideMenue extends StatefulWidget {
   @override
-  _SideBarState createState() => _SideBarState();
+  _SideMenueState createState() => _SideMenueState();
 }
 
-class _SideBarState extends State<SideBar>
-    with SingleTickerProviderStateMixin<SideBar> {
-  AnimationController _animationController;
-  StreamController<bool> isSidebarOpenedStreamController;
-  Stream<bool> isSidebarOpenedStream;
-  StreamSink<bool> isSidebarOpenedSink;
-  final _animationDuration = const Duration(milliseconds: 500);
-
-  @override
-  void initState() {
-    super.initState();
-    _animationController =
-        AnimationController(vsync: this, duration: _animationDuration);
-    isSidebarOpenedStreamController = PublishSubject<bool>();
-    isSidebarOpenedStream = isSidebarOpenedStreamController.stream;
-    isSidebarOpenedSink = isSidebarOpenedStreamController.sink;
-  }
-
-  @override
-  void dispose() {
-    _animationController.dispose();
-    isSidebarOpenedStreamController.close();
-    isSidebarOpenedSink.close();
-    super.dispose();
-  }
-
-  void onIconPressed() {
-    final animationStatus = _animationController.status;
-    final isAnimationCompleted = animationStatus == AnimationStatus.completed;
-
-    if (isAnimationCompleted) {
-      isSidebarOpenedSink.add(false);
-      _animationController.reverse();
-    } else {
-      isSidebarOpenedSink.add(true);
-      _animationController.forward();
-    }
-  }
-
+class _SideMenueState extends State<SideMenue> {
   @override
   Widget build(BuildContext context) {
-    final screenWidth = MediaQuery.of(context).size.width;
+    return Container(
+      width: 200,
+      child: Drawer(
+        child: Column(
+          children: [
+            SizedBox(
+              height: 50,
+            ),
 
-    return StreamBuilder<bool>(
-      initialData: false,
-      stream: isSidebarOpenedStream,
-      builder: (context, isSideBarOpenedAsync) {
-        return AnimatedPositioned(
-          duration: _animationDuration,
-          top: 0,
-          bottom: 0,
-          left: isSideBarOpenedAsync.data ? 0 : -screenWidth,
-          right: isSideBarOpenedAsync.data ? 0 : screenWidth - 45,
-          child: Row(
-            children: <Widget>[
-              Expanded(
-                child: Container(
-                  padding: const EdgeInsets.symmetric(horizontal: 20),
-                  color: const Color(0xFF4A4A58),
-                  child: Column(
-                    children: <Widget>[
-                      SizedBox(
-                        height: 40,
-                      ),
-                      ListTile(
-                        title: Text(
-                          "Glossary",
-                          style: TextStyle(
-                              color: Colors.white,
-                              fontSize: 25,
-                              fontWeight: FontWeight.w800),
-                        ),
-                        subtitle: Text(
-                          "Faculty of Applied Science",
-                          style: TextStyle(
-                            color: Colors.blueAccent,
-                            fontSize: 12,
-                          ),
-                        ),
-                        leading: CircleAvatar(
-                          child: Image.asset('assets/logo.png'),
-                          radius: 30,
-                        ),
-                      ),
-                      Divider(
-                        height: 20,
-                        thickness: 0.5,
-                        color: Colors.white.withOpacity(0.3),
-                        indent: 32,
-                        endIndent: 32,
-                      ),
-                      MenuItem(
-                        icon: Icons.radio_button_unchecked,
-                        title: "CIS",
-                        onTap: () {
-                          onIconPressed();
-                          BlocProvider.of<NavigationBloc>(context)
-                              .add(NavigationEvents.CISWordsClickedEvent);
-                        },
-                      ),
-                      MenuItem(
-                        icon: Icons.radio_button_unchecked,
-                        title: "PST",
-                        onTap: () {
-                          onIconPressed();
-                          BlocProvider.of<NavigationBloc>(context)
-                              .add(NavigationEvents.PSTWordsClickedEvent);
-                        },
-                      ),
-                      MenuItem(
-                        icon: Icons.radio_button_unchecked,
-                        title: "NR",
-                        onTap: () {
-                          onIconPressed();
-                          BlocProvider.of<NavigationBloc>(context)
-                              .add(NavigationEvents.NRWordsClickedEvent);
-                        },
-                      ),
-                      MenuItem(
-                        icon: Icons.radio_button_unchecked,
-                        title: "FST",
-                        onTap: () {
-                          onIconPressed();
-                          BlocProvider.of<NavigationBloc>(context)
-                              .add(NavigationEvents.FSTWordsClickedEvent);
-                        },
-                      ),
-                      MenuItem(
-                        icon: Icons.radio_button_unchecked,
-                        title: "SSPE",
-                        onTap: () {
-                          onIconPressed();
-                          BlocProvider.of<NavigationBloc>(context)
-                              .add(NavigationEvents.SSPEWordsClickedEvent);
-                        },
-                      ),
-                      Align(
-                        alignment: Alignment.bottomLeft,
-                        child: Text(
-                          'Created by @Hansajith',
-                          style: TextStyle(
-                              color: Color(0xFF262A60), fontSize: 10.0),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
+            // Faculty of Applied sciences
+
+            ExpansionTile(title: Text('Applied sciences'), children: [
+              ListTile(
+                title: Text('CIS'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.CISWordsClickedEvent);
+                  Navigator.pop(context);
+                },
               ),
-              Align(
-                alignment: Alignment(0, -0.9),
-                child: GestureDetector(
-                  onTap: () {
-                    onIconPressed();
-                  },
-                  child: ClipPath(
-                    clipper: CustomMenuClipper(),
-                    child: Container(
-                      width: 35,
-                      height: 110,
-                      color: Color(0xFF4A4A58),
-                      alignment: Alignment.centerLeft,
-                      child: AnimatedIcon(
-                        progress: _animationController.view,
-                        icon: AnimatedIcons.menu_close,
-                        color: Color(0xFF1BB5FD),
-                        size: 25,
-                      ),
-                    ),
-                  ),
-                ),
+              ListTile(
+                title: Text('PST'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.PSTWordsClickedEvent);
+                  Navigator.pop(context);
+                },
               ),
-            ],
-          ),
-        );
-      },
+              ListTile(
+                title: Text('NR'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.NRWordsClickedEvent);
+                  Navigator.pop(context);
+                },
+              ),
+              ListTile(
+                title: Text('FST'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.FSTWordsClickedEvent);
+                  Navigator.pop(context);
+                },
+              ),
+              ListTile(
+                title: Text('SSPE'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.SSPEWordsClickedEvent);
+                  Navigator.pop(context);
+                },
+              ),
+            ]),
+
+            // Faculty of Agriculture
+
+            ExpansionTile(title: Text('Agriculture'), children: [
+              ListTile(
+                title: Text('Agri business managment'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.CISWordsClickedEvent);
+                  Navigator.pop(context);
+                },
+              ),
+              ListTile(
+                title: Text('Exprt agri culture'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.PSTWordsClickedEvent);
+                  Navigator.pop(context);
+                },
+              ),
+              ListTile(
+                title: Text('Live stock production'),
+                onTap: () {
+                  BlocProvider.of<NavigationBloc>(context)
+                      .add(NavigationEvents.NRWordsClickedEvent);
+                  Navigator.pop(context);
+                },
+              ),
+            ]),
+          ],
+        ),
+      ),
     );
-  }
-}
-
-class CustomMenuClipper extends CustomClipper<Path> {
-  @override
-  Path getClip(Size size) {
-    Paint paint = Paint();
-    paint.color = Colors.white;
-
-    final width = size.width;
-    final height = size.height;
-
-    Path path = Path();
-    path.moveTo(0, 0);
-    path.quadraticBezierTo(0, 8, 10, 16);
-    path.quadraticBezierTo(width - 1, height / 2 - 20, width, height / 2);
-    path.quadraticBezierTo(width + 1, height / 2 + 20, 10, height - 16);
-    path.quadraticBezierTo(0, height - 8, 0, height);
-    path.close();
-    return path;
-  }
-
-  @override
-  bool shouldReclip(CustomClipper<Path> oldClipper) {
-    return true;
   }
 }
